refactor: extract shared listingSource and useListingActions hook
- Move duplicated listingSource() from ListingCardBody and
  ListingDetailPage to shared lib/utils.ts
- Create useListingActions hook that encapsulates the bookmark/seen
  optimistic toggle pattern, eliminating ~100 lines of duplication
  between ListingsPage and ListingDetailPage

closes #1097
closes #1098
